LayoutTests/js/intl-expected.txt

This test checks the behavior of the Intl object as described in 8 of the ECMAScript Internationalization API Specification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Intl is an instance of Object
PASS typeof Intl is 'object'
PASS Object.prototype.toString.call(Intl) is '[object Object]'
PASS Object.getPrototypeOf(Intl) === Object.prototype is true
PASS new Intl threw exception TypeError: Object is not a constructor (evaluating 'new Intl').
PASS Intl() threw exception TypeError: Intl is not a function. (In 'Intl()', 'Intl' is an instance of Object).
PASS delete Intl; is true
PASS 'Intl' in global() is false
PASS successfullyParsed is true

TEST COMPLETE
